💄 style(chat): move completion error above input trays
